pyproject.toml
[tool.poetry]
name = "gcp-backup-exercise"
version = "0.1.0"
description = ""
authors = ["Justin Baldwin"]
readme = "README.md"
packages = [{include = "gcp_backup_exercise"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
google-cloud-pubsub = "^2.15.0"
google-cloud-resource-manager = "^1.8.1"
google-cloud-bigquery = "^3.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
